feat(#573): split responsibilities
